pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "realtime-agentic-api"
version = "0.1.0"
description = "Realtime Agentic API - Serverless event-driven AI agent platform on AWS"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
    "pydantic>=2.10.0",
    "boto3>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "hypothesis>=6.100.0",
    "moto>=5.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
]

[tool.setuptools.packages.find]
include = ["infra*", "runtime*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "property: property-based tests (Hypothesis)",
    "unit: unit tests",
    "integration: integration tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
